added upload, refactored routes
